third_party/WebKit/LayoutTests/platform/win/external/wpt/shadow-dom/untriaged/styles/test-003-expected.txt

This is a testharness.js-based test.
PASS A_06_00_03_T01 
PASS A_06_00_03_T02 
FAIL A_06_00_03_T03 assert_equals: Style sheet is not accessible via styleSheets expected 1 but got 0
Harness: the test ran to completion.
